Remove insecure sha512 password hashing
This also switches to use the high-level passlib library for all password
hashing. This improves security and provides mechanisms to seamlessly
migrate user databases to use other hash algorithms (like Argon2) in the
future.
