+++ girder/conf/girder.dist.cfg
@@ -3,14 +3,6 @@ server.socket_host = "127.0.0.1"
 server.socket_port = 8080
 server.thread_pool = 100
 
-[auth]
-# Use "bcrypt" or "sha512"
-hash_alg = "bcrypt"
-
-# Exponent for bcrypt complexity (default=12).
-# Don't change this unless you know what it means.
-bcrypt_rounds = 12
-
 [database]
 uri = "mongodb://localhost:27017/girder"
 replica_set = None
