+++ girder/utility/model_importer.py
@@ -9,7 +9,7 @@ def _registerCoreModels():
         return
 
     from girder.models import (
-        api_key, assetstore, collection, file, folder, group, item, notification, password,
+        api_key, assetstore, collection, file, folder, group, item, notification,
         setting, token, upload, user)
 
     ModelImporter.registerModel('api_key', api_key.ApiKey)
@@ -20,7 +20,6 @@ def _registerCoreModels():
     ModelImporter.registerModel('group', group.Group)
     ModelImporter.registerModel('item', item.Item)
     ModelImporter.registerModel('notification', notification.Notification)
-    ModelImporter.registerModel('password', password.Password)
     ModelImporter.registerModel('setting', setting.Setting)
     ModelImporter.registerModel('token', token.Token)
     ModelImporter.registerModel('upload', upload.Upload)
